chore: fix rewrites and remove /docs prefix in docs website
