+++ apps/docs/content/docs/openapi/meta.json
@@ -7,6 +7,9 @@
     "index",
     "---Application---",
     "AppController_getHello",
+    "AppController_getPlannerData",
+    "AppController_getPlannerEntities",
+    "AppController_getPlannerEntity",
     "AppController_getShardsEmbed",
     "AppController_getTimesEmbed",
     "---Bot Statistics---",
