+++ apps/docs/middleware.ts
@@ -1,0 +1,9 @@
+import { NextResponse, type NextRequest } from "next/server";
+
+export function middleware(request: NextRequest) {
+  return NextResponse.redirect(new URL("/docs", request.url));
+}
+
+export const config = {
+  matcher: ["/"],
+};
